Delete duplicated product in case of error only if it exists

+++ src/Adapter/Product/AdminProductDataUpdater.php
@@ -233,7 +233,9 @@ class AdminProductDataUpdater implements ProductInterface
                 return $product->id;
             }
         } else {
-            $product->delete();
+            if ($product->id !== null) {
+                $product->delete();
+            }
             throw new \Exception('An error occurred while creating an object.', 5009);
         }
     }
